--- src/components/Board/Entry/index.ts
@@ -1,2 +1,0 @@
-export { BoardEntry } from "./BoardEntry";
-export type { IBoardEntry } from "./BoardEntry";
