+++ src/components/Leaderboard/mockData.json
@@ -1,0 +1,12 @@
+[
+  {
+    "name": "Дэн",
+    "avatar": null,
+    "score": 1500
+  },
+  {
+    "name": "Иван",
+    "avatar": null,
+    "score": 900
+  }
+]
