--- src/components/Leaderboard/mockData.ts
@@ -1,16 +1,0 @@
-import { IBoardEntry } from "components/Board/Entry";
-
-const mockData: IBoardEntry[] = [
-  {
-    name: "Дэн",
-    avatar: null,
-    score: 1500,
-  },
-  {
-    name: "Иван",
-    avatar: null,
-    score: 900,
-  },
-];
-
-export default mockData;
