+++ src/components/UI/Layout/Layout.tsx
@@ -2,11 +2,9 @@ import React, { FC } from "react";
 
 import { Header } from "../Header";
 
-export const Layout: FC = ({ children }) => {
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-    </>
-  );
-};
+export const Layout: FC = (props) => (
+  <>
+    <Header />
+    <main>{props.children}</main>
+  </>
+);
